(redo) Replace gradle APIs with project system APIs.
Replace GradleSyncListener API calls in waitForGradleProjectSyncToFinish
and use project system's sync manager APIs instead.

Note that this CL keeps the existing behaviour of
waitForGradleProjectSyncToFinish and adds documentation for the race
condition it introduces where the method will not return if it is called
after the sync that it's supposed to wait for completes (b/70937990).
This CL does not fix the bug.

This CL is a resubmission of a previous CL that broke several ui-tests
due to gradle sync reporting failures upon PARTIAL_SUCCESS, which
ui-tests don't treat as failures. The error detection logic now checks
for both successful sync as well as SyncResult.SOURCE_GENERATION_FAILURE
to make sure a sync still passes if the sync was successful but the
build following the sync failed.

Implementation of waitForGradleImportProjectSync has been changed to use
project system's sync manager API.

Test: Manually verified that tests under groups {PROJECT_SUPPORT,PROJECT_WIZARD,THEME,EDITING,DEFAULT} that passed before this change still pass.
Change-Id: Ic51e5e16693bfd0ef8a058c51e01f52c4f792e6f