Delete obsolete method AndroidFacet.androidPlatformChanged
With the recent changes in If90eea6b448a2 and Ie48c058777bad,
there is no need for the facet to listen to project root changes
anymore.

AndroidFacetConfiguration explicitly invokes androidPlatformChanged()
whenever the facet configuration is changed from the project structure
dialog (this only applies to JPS projects). This has been migrated over
to indicate a facet configuration change instead, which is subscribed to
by whoever is interested.

Fixes: 71553467
Test: code deleted, no new functionality, existing tests pass
Change-Id: I8baf59bf99466de69e3ced75fa01790d463c0f8b