Add library dependency with different scopes.
When adding library dependencies, we check if the library
has been added to current module, if it does, then won't
add it again.

When checking if one library has been added, current code
only checks that the library with the same name exists or
not.
But the same library dependency can have different scopes.

We should also compare the scopes, and if an existing libray
is added with a different scope, then a new order entry should
be created with the new scope.

Bug: 71452960
Test: testSetUpLibraryTwiceWithSameLibraryInDifferentScopes.
Change-Id: I93b10e5ee6dbc6a2a61b85747060d331599ca4bc