Move the creation of template menu out of AndroidFacet
The template menu (a contextual menu that shows up when you right click
on folders and allows the creation of various new android components)
needs to be created just once per Android project. This CL moves it to
the AndroidProjectComponent.

Unfortunately, the flag that ensures that this creation only happens
once still remains: it turns out that this menu creation needs to
happen once in the entire application, and not once per project. In
order to achieve that, we'll need to put this in an
ApplicationComponent. However, I'm not performing that in this CL
because the whole menu creation itself seems somewhat busted (it is
for instance refreshed on every gradle sync) and might need a larger
rethink, which is out of scope of the current CL whose primary objective
is to simplify the AndroidFacet.

Bug: 71553525
Test: existing UI tests pass
Change-Id: Ibaed771cea1e6693bb2a79c8938d6d81fadee19d