Ensure that only Gradle files belonging to the project are processed.
This change adds a check to make sure that the GradleFileChangeListener
in GradleFiles only responds to events from files inside its own
project.

This should be taken care of by the PsiManager but it appears there are
instances where the listener gets triggered for events in files that
are not part of its own project.

Bug: 71815594
Test: N/A
Change-Id: Ia365446af56012c905d5eb26e70300d1578f280c