Fixing NPE after AddExistingDestinationMenuTest
In AddExistingDestinationMenuTest.testNewComponentSelected, we started
the scroll-to-center executer, which runs events every 10 ms for a
little bit. If the test ends during that time, the NavigationSchema
would be disposed and subsequent events would NPE when trying to get it.
Fixed NavDesignSurface to cancel the job when it's disposed. Also made
NavigationSchema run its dispose later, so any existing events would go
first.

Test: should fix tests shown as failed because of this exception
Change-Id: Ibb1c78e5fa6c0cc47af490f575130f38f1191177