Properly track clients in sessions for IR
AndroidProcessHandler code has the incorrect assumption that
applicationId will always match the client's description.

This assumption is false if the app uses android:process attribute on
the default activity. Then the description is:

in the case of a local process: $packageId:$process, we should match
against packageId. This CL handles that.

in the case of a global process: $process, that requires a bigger
change in a follow up CL to parse the manifest to determine the process
of the launching activity.

Bug: b/71645350
Test: TODO
Change-Id: I2c83ace8bf5e0b0358150559594d6d592abf767a
(cherry picked from commit 2a1edf8ceb1e1235fea48c0d995b4a63c623e28e)