Enable output parsers on GradleBuildInvoker, add a AGPBI parser.
BuildOutputInstantReaderImpl was created during a try with resource  since
it implements a closable, However, this class creates a thread that should
be kept runing even after createBuildTaskListener is running, but since its
close method is called as soon as the try finishes, then none of the parsers
are called. This causes to not report any errors from them in the Build
window.

This CL removes the try with resource to cause close to be called inside
ExternalSystemTaskNotificationListenerAdapter#onEnd and also adds a new type
of parser for AGPBI messages.

Test: New test for AGPBIOutputParser
Bug: 71845632
Bug: 70299236

This is how the generated messages look like (they are navigable):

https://screenshot.googleplex.com/rRVHTYbehiM

Change-Id: I159fa04b69141704fb78f90d3b851fb0dafe8b7c
(cherry picked from commit 034b3b3df532caff79564c020ad96f59a3d03a9a)