Disable Apply Changes button during empty sessions
The IR button should be disabled if the AndroidSessionInfo contains no
devices. There is not target for IR to deploy against.

This is one of the causes for the bug. As the client fails to be added
to the AndroidSessionInfo when using a global process name. So the Apply
Changes button shouldn't be enabled in the first place.

Fixes: b/71645350
Test: n/a
Change-Id: I4912742f13790fbdd402471998a563d755fe0279
(cherry picked from commit 2b79ce0eb711d68d0a591ccd4a5ee78e9b4123e1)