Remove the getAndroidModel accessor from AndroidFacet
Callers will now get the model directly on the configuration object.
This change was produced by the inline refactoring.

Added noinspection comment to AndroidXmlSpellcheckingStrategy
to suppress bogus linter error that is blocking the PSQ.

Removed problematic and unnecessary mocking from
AndroidRunConfigurationBaseCanInstantRunTest and
GradleApkProviderGetApkTest.

Rewrote the mocks in NewAndroidComponentActionTest to conform
to the new API.

Bug: 71553325
Test: n/a
Change-Id: I62d2dd34d8d5f764a77c820c1f23a90ce0c6a139