Prevent inactive steps from affecting the current step
Wizard steps are only allowed to access a very small subset of their
parent wizard's functionality, but it turns out even then, that
some steps were accidentally triggered functionality even when they
weren't yet active. For example, when creating a new project, the
second step would tell the parent wizard to refresh its state
(not intentionally but as a side-effect for code it was setting
up), which caused the first step's "Next" button to flicker.

This, in turn, would cause the NPW UI tests to flake, albeit
(though perhaps worse) rarely.

Fixes: 71918418
Test: NewProjectTest
Change-Id: I4a87216f37a6e2f3694f2aa5bc5e8d1a8be8a1a6