Fix race condition in dual debugger tests
The IDE creates two debug tabs in the debugger window for dual
debugger tests. The first tab is named "app", and this tab has a
container for printing values of variables in native code. The second
tab is named "app-java", and it contains a container for printing
values of variables in Java code.

Description of test flow:

The test sets two breakpoints: one breakpoint in native code, and one
breakpoint in Java code. The dual debugger is used for the debug
sessions.

Tests check for when breakpoints are hit by searching for a GridImpl
object.  Tests conclude a breakpoint has been hit if the GridImpl
container contains a set of expected variables and the expected
variables' values. GridImpl instances are JPanel instances where
variables and their values in an application being debugged are
printed.

In the test, the native code breakpoint is hit first. The "app" tab is
activated, and the test checks the "app" debug window for a GridImpl
object. If the expected variables and their values are shown on the
screen, the test concludes that the native code breakpoint has been
triggered, and the test continues by clicking on the resume button.

After hitting the resume button, the Java breakpoint is triggered.
The test checks if the Java breakpoint has been triggered by searching
for a GridImpl object and querying its contents for variables and
their values. We expect the variables and their values will match the
expected values given in the test, and this is the end of a passing
test.

Description of race condition:

Suppose the test has resumed the application after the native code
breakpoint was hit. The active tab is still the "app" tab. The test
starts polling for GridImpl containers to check if the Java breakpoint
has been hit.

Assume the test retrieves the GridImpl instance from the "app" tab.

Before the test queries this GridImpl object, the IDE may switch the
active debugger tab to the "app-java" tab because the Java breakpoint
was hit. In this situation, the test will have a reference to the
native debugger GridImpl container that's no longer shown on the
screen, and the test will wait until the container is displayed again.
However, the container will never show up on the screen again because
the "app-java" tab is shown, so the test will fail.

There is an assumption of atomicity between the retrieval of the
GridImpl instance and the query of the GridImpl instance. This is not
the case, so test writers should be more careful when working with
dual debuggers.

The solution here is to wait for the "app-java" tab to be activated
before searching for the GridImpl object. A wait was added to
DebuggerTestBase.checkAppIsPaused to wait for the given tab - in this
case, the "app-java" tab - to be active.

Change-Id: Ib8a12af7c1f78039aa4eaa52e09ebfb0108ed281
Fixes: 72163607, 72164324
Bug: 72163607, 72164324
Test: BasicNativeDebuggerTest.testDualDebuggerBreakpoints