Parameterize a UI test by target build system
This CL enables NlEditor.basicLayoutEdit to be parameterized by the
build system (Gradle + Bazel).

The existing GuiTestRunner now takes in an optional constructor
parameter indicating the build system to be used in a test. The
GuiTestRunner uses this build system for two purposes:
 1. It only tests methods in a class that are annotated to work with the
 given build system. (Although it assumes that no annotations => works
 only with Gradle).
 2. When instantiating a test, it passes along the build system to use
 via the IdeTestApplication instance, from where the GuiTestRule can
 read that setting.

The GuiTestRule has been modified slightly: It now delegates its build
system specific work to the appropriate GuiTestSystem.

The test itself is parameterized using the regular Junit4
parameterization annotation, with the various build systems as the
parameters. It is further annotated with a specific test runner factory
(GuiTestRunnerFactory), which instantiates the GuiTestRunner with the
correct build systems.

All existing tests that are not parameterized continue to work as usual
sine the GuiTestRunner defaults to Gradle build system.

Test: updated
Test-Strategy: ui.test.groups=DEFAULT
Fixes: 70400243
Change-Id: I75c56ef97441e8d45e39f364dbda1fa6e92baaf0