Include build-system information in test runner name.
Currently when running a ui-test with multiple TargetBuildSystems, the
tests show up twice with the exact same name.  This is confusing as the
user wouldn't know which build system the test was running with unless
they parse through runtime logs.

This CL adds better naming by providing a better name from the
GuiTestRunner.  With the test runners named this way, the tests are
split under separate run branches in the UI with one per build system.

Test: Nothing broke.
Change-Id: I8c6f3b562898fe96bde249dafa68741e2e45e569