Infinite loop can occur in cleanup flow
When a modal dialog is closed with a close-window event, the dialog can
open another instance of itself again. This causes the
checkForModalDialogs method to end up in an infinite loop, since there
will always be a dialog open.

With this change, EspressoRecorderTest is now safe to run again.

Test: EspressoRecorderTest, ran locally. Infinite loop was no longer
infinite
Change-Id: I40bc73646a2113a49298cacf68567259e3ffc8e0