Remove call to abstract method from constructor
In SceneManager, the constructor was calling an abstract method which
could then call on some properties of the child class before they have
been initialized (as it is called in the super constructor for the
children classes).
This moves the call to this abstract method outside of the constructor
of the parent and into the constructor of each child.

Bug: 72574190
Test: NavigationViewTest
Change-Id: I2cc225c308d7d20972cab183ceb0bbb34c047eff