project already disposed in NlDesignSurface
This follows Change-Id: Idd56095267294c4f9dce65624ba61eca4cc2ec62
The Layout Editor Surface should stop rendering when the current
project is closed. However, it is possible that it makes a call to
GradleProjectInfo.isBuildWithGradle/Blaze.isBlazeProject, etc when
the project is still not disposed, and because some (gradle)
isBuildWithGradle runs inside a read lock, by the time
isBuildWithGradle() is executed the project is disposed.
This CL now wraps the call to create the RenderErrorModel inside a
ReadAction and only then checks for project disposed/open.

Change-Id: I961a9ec7b20f7a6e84a1524f767d42d5d8cfc84e
Fixes: 70682962
Test-Strategy: ui.test.groups=DEFAULT
Test: Run NlEditorTest