[Gradle Build Model] Support vars from applied files.
The changes adds support for applied Gradle build files that are in the
basic format "apply from: "file.gradle"". In order to do this, when the
file is applied we merge all of the resulting properties into the
current file being parsed.

We add a new ElementState (APPLIED) to the GradlePropertiesDslElement to
indicate that these merged properties shouldn't be modified by this file.

We also ensure that when a GradleDslFile is updated by calling its
apply() method, that all its applied files are updated first.

Bug: 72432440
Test: Unit test added.
Change-Id: I290a3aa4ebbfaeddd9c0f26b540dbccaf953b25b