NavSetStartProviderTest cleans its model objects
Earlier, the test class had a lateinit member variable 'model' and
it was initialized in setUp.  But this caused the test runner to hold
onto the model and project created by every test in the class.

A better way is to make the model local to every test.  It ensures
proper clean-up after every test.

Another solution could have been to make 'model' nullable and set it to
null in tearDown.  The previous solution, however, is less number of
lines, and looks good enough for this class with only two test methods.

Test: YourKit memory profiling
Bug: 72755830
Change-Id: Ic2863958985ec97f60c28364386dc75e472fb3e8