[ScreenRecord] Check which recording scheme to use before using.
The condition was only checked when determining whether to enable the
recording button, but not when we were about to record, which may be
different if the device in question changed.

BUG: b/73005414

Test: Record with emulator first (with 'ScreenRecording' adv feature
enabled), then record with real device. The first should be in .webm
format and the latter in .mp4 format.

Change-Id: Iea5c70a97de147e36508700c71daa4791c059e50