Make findConsoleView return only visible consoles.
GradleTasksTest was disabled because it was timing out. The reason of
the timeout is that the implementation of ContentFixture#getOutput
depends on findConsoleView, if two or more ConsoleViewImpl are on the same
window, then this function returns null, which causes getOutput to spin.

With this change, we only look for the consoles that are currently visible
(only one can be shown at any time).

findConsoleView is only used for getOutput and for clickClearAllButton
and in both cases it is done while the console is visible.

Test: Disabled tests are enabled after this change.
Bug: 70684215
Change-Id: I61450c42465e812c67cdb85fb4d6d47f374776c9