Create a new BuildOutputInstantReader when a build is restarted.
When GradleBuildInvoker#createBuildTaskListener is called, an instance
of BuildOutputInstantReaderImpl is used to process the build output and
once the build is finished, the reader must be closed. The onStart event
on the listener creates a StartBuildEvent with a restart action that also
uses the same reader. Since the reader was close once the build has
finished, the restart action output can not processed by the reader.

This CL creates a new reader when the restart action is performed.

Bug: 73102585
Test: Manually tested, need to consider for UI test.
Change-Id: Id25116c61549f1e8f17876aca5f8a640f60e3ef5