Do not forget APK signing key password stored with a legacy requestor.
This CL is essentially a follow-up for the fix submitted earlier with
ag/3265044. That CL correctly fixed the problem with APK signing key store
password being forgotten even when 'Remember passwords' was checked.
Also, that CL made an effort attempted to ensure smooth transition
from the Studio versions which don't contain that fix to the versions
which do. For that it tried to request APK passwords using in turn both
the new and legacy requestors, thus the user wouldn't have to retype their
key password - only the keystore password once, which was the subject of the
original bug.

However that fix itself contained a typo in the classname
used as a legacy requestor, thus resulting in returned 'legacy' passwords
always being null. Therefore those passwords were never populated,
defeating the effort to ensure smooth transition to the fixed versions.
The user would only need to retype both the key and keystore passwords in
order to get them stored correctly (that part of the fix was working as
expected).

Therefore this CL addresses that problem by correcting the requestor
classname, and adding a dedicated test for it. Also this CL improves the test
infrastructure for the KeystoreStep by adding PasswordSafeMock - this allows
replace the PasswordSafe app service for each test and not run into side
effects of the previous tests.

Bug: b/70937387
Test: Added KeystoreStep.testRememberPasswordsUsingLegacyRequestor
Change-Id: Idfe599b10061728c8cd95ef8d1cd0c741a3babe6