Move energy logic into the datastore
This CL introduces the datastore layer for the energy profiler,
moving the "fake data" logic that currently exists in EventService into
EventDataPoller.

This is still a somewhat temporary situation, as in a followup CL, events
will be fetched from the device instead, and the logic for samples will be
updated to real data.

Fix: 73081674
Bug: 73116415

Test: Tests added, except for EnergyDataPoller. EnergyDataPoller is
      currently generating fake, random data. Will add tests in
      followup aCLs once real data is supported.

Change-Id: I3f9586f22b4462ddbf4c6fd87fff50b3e047574c