Use paths relative to project root directory in Bazel UI tests
Our existing UI tests assume that the top-level project directory
(the one that contains all source files related to a project) and
the project base directory (the one containing .idea/ ) are the same.
This is true when targeting a Gradle project, but not for Bazel
projects, for which the import project wizard automatically places
.idea/ inside a .project/ directory.

This CL introduces the getProjectRootDirectory() method in
GuiTestProjectSystem, providing a build system-independent way
for IdeFrameFixture to locate files relative to the project root
directory.

The Bazel UI tests that were disabled due to this issue have been
re-enabled.

Fixes: 72952881
Test-Strategy: ui.test.groups=DEFAULT
Test: All re-enabled Bazel UI tests pass. NlEditorTest#basicLayoutEdit
      still passes when targeting a Gradle project.

Change-Id: Ie6869787b589ad7b9ae80581a9cc4e9b32f2c693