Support for session selection
Added selection support in the sessions panel, and also mechanism
to support rendering different session artifacts (e.g. Session, hprof,
CPU capture) in the panel.

This change also introduced the concept of a profiling session which
can be different from the currently selected session. This allows
users to browse around historical session without having to stop profiling
a session that that is in progress

Minor tweaks:
- removed "SESSION" label when the panel is collapsed
- ensure the expand/collapse button states are correct when the panel
  is toggled
- updated ProfilerTimeline to do a more accurate update after reset.

Bug: 67509537
Bug: 67526434
Test: added and updated
Change-Id: Ibd656830d862fd68f22e65178902589c280753d8