Make sessions work without relying on device info
This concerns MemoryProfiler only, as it previously relies
on the device's info to determine whether JVMTI-based allocation
tracking is used. With this change, the profiler now looks at both
the SessionMetaData and whether there is a non-legacy AllocationsInfo
sample within a session's time range to determine if JVMTI tracking
is used.

Bug: 73652143
Test: Updated

Change-Id: I2b39a2f85952faa0b82b6aa49702065916eb4d34