[Gradle Property Model] Add transform support for complex properties.
This change moves three of the properties in SigningConfigModel to use
the GradlePropertyModel. As a prerequisite for this we have added
transform support to the GradlePropertyModel.

These transform allow us to provide a set of functions that map from the
GradleDslElement on the property to the one we actually care about.
For example, a typical fileStore property:
  fileStore file('file.txt')
Now since we only have the capacity to represent simple types we define
a mapping between the whole property element and the element
representing "file.txt". These transforms will also be used for
properties with different supported type (such as storePassword) as then
enable us to convert values to different forms.

We also create a GradlePropertyModelBuilder to help with the
construction of GradlePropertModels.

Bug: 73184484
Test: Unit tests updated
Change-Id: I22ecd60297fc82ef3fb588d7d7a561faedd48af4