Implement toJavaCode in PsiModelClass for Binding
This CL fixes a crash where we would try to generate the java code
for a PsiModel class. Code generation is not supported in IJ and
this was happening as a side effect of checking equals in the base
class.

Since it is relatively easy to generate class representation via
PsiType, this CL makes PsiModelClass implement tojavaCode().

Bug: 73644073
Test: DataBindingTest#resolveXmlReferences
Change-Id: Ie5ea5b3e11055222c3d13ae9a3adf8c5e6cca145