Ensures CPU captures are within range once selected
Add an API to ProfilerTimeline to make sure a given range totally fits
the timeline view range. To achieve that, we zoom out until the view
range is big enough to fit the target, then jump to the target to make
sure it's within the view range.

When selecting a CPU capture, we will call the method added to
ProfilerTimeline to make sure the capture fits within the view range (with
a padding of 5% of the view range) and is entirely displayed.

Bug: 73338399
Test: ProfilerTimelineTest#jumpToTargetOnTheLeft,
      ProfilerTimelineTest#jumpToTargetOnTheRight, and
      ProfilerTimelineTest#jumpToTargetWithinViewRange

Change-Id: I4d5459850fd7b0ab307241bad055ed2ef50a90cc