Fix invalidation of PsiResourceItems
When getting out of the dumb mode, PsiResourceItems inner XmlTag might
be invalidated. Since they might be cached, this will cause crashes
along the way.
This fixes that invalidation by moving to use SmartPointers so the
XmlTag can still be correctly retrieved after the file invalidation
since the file is still there, just invalid after the dumb mode.
A lot of the incrementality of the ResourceFolderRepository is based on
detecting changes based on the pointer to the XmlTag so added a new
method to check if the item was pointing at some point to that tag.

Fixes: 73623886
Test: Added testFileInvalidationAfterDumbMode
Change-Id: Iad1cdfbcae645ad5fe0a3eb4d274122b753801cd