Log issues in ToolWindowFixture without NPEs
There are cases where the project pane tool window never appears, even
though screenshots and local test runs suggest that it is visible.
This change logs some information about the UI when the tool window is
not found.

This is a re-implementation of
I8a8fb13724eb44d288c651171e134483fc3b4fa5.

Change-Id: I238cd87bfafa08d59c803d81a0ba0d2a9466c5ee
Bug: 72042463
Test: ran DependenciesTest locally. Sanity tests in post-submit