[Gradle Build Model] Add MultiTypePropertyModel
This change adds a new type of property model that can be used to
represent properties with more than on type. The type and value of these
properties can then be changed.

This is required in order to create a model for dealing with passwords
where we would like to be able to set multiple forms.

We also add some missing javadocs in PropertyTransform.java

Bug: 73184484
Test: Unit tests will be in a following change
Change-Id: I4a7e4216d1ea65209464cb7ff529561b057324e7