Revert "Gather information to why AVD is not created"
This reverts commit 0a84f675c5fbd1057687d9acc25b2d0f73c9e6e6.

Reason for revert: Found underlying cause for why this was happening. Copied from bug 72748687, the issue was a race condition in the code used to build the UI for the deploy target picker dialog. There is a possibility where one block of code can end up overwriting a list model reference with a reference to another list model that has no AVDs to display. The end result is that the dialog can end up showing no AVDs. 

Moreover, logs from our CI system definitely show that AVD files are being deleted at the end of the our test when this bug occurs, which is strong evidence that the AVD is being created.

Bug: 72748687
Change-Id: I4db039c8e87600fd0e7e8d2c769f9cf77a36a5e5