Check if project is initialized in MissingAndroidPluginErrorHandler
When a project is imported and it contains an AGP version that cannot be
found, an error message is generated. This is processed by
MissingAndroidPluginErrorHandler, but if the project is not yet initialized,
a Logger$EmptyThrowable is generated when trying to see if google repository
is present.

This CL solves this issue by first checking if the project is initialized
and act accordingly. If the project is initialized, we can check and
decide if we show the quickfix. If the project is not initialized, we add
the quickfix with a slightly modified message.

Bug: 73948080
Test: New test to confirm the correct hyperlinks are created when project is
not initialized.

Change-Id: I4c91f2613f3b35d539907ba61df46ce8097bfff2
(cherry picked from commit c65f219318f9087e979217dffa42a85039d17edd)