[New PSD] Fix adding multiple dependencies without syncing.
In this CL we start collecting module dependencies from both parsed and
resolved models. Thus adding a new dependency has now been reduced to simple
adding the dependency to the parsed model and recollecting the dependencies.

The CL temporary disables several tests that rely on proper matching of
parsed and resolved dependencies. They will be re-enabled once matching is
reworked.

Fixes: 73886954
Test: unit tests uncommented
Change-Id: Ice84ec5543970a93b4048d90126864bb10cbefba