Try gracefully killing emulator process again
This reverts commit 65e7d5f391eb4324d767fbf78a968fbc0be6403b. There were
test failure results that showed the emulator state was never updated
to DISCONNECTED. There are two possibilities:

1. The emulator never terminated.
2. There was a thread issue, where the variable representing the device
state was not visible to other threads when it was updated.

This change addresses the second possibility. Rather than checking the
state of the IDevice object, the emulator-kill code retrieves the list
of known devices connected to ADB. The data structure containing the
IDevice objects is a thread-safe CopyOnWriteArrayList.

If tests still fail with an error message suggesting that the emulator
failed to terminate, then the failures are caused by the emulator
failing to terminate rather than being caused by variable visibility
issues to other threads.

Test: all existing emulator tests
Bug: 74035584
Change-Id: I0fc00af807591a651892f5ec406f88d8722646ee