[Gradle Build Model] Add project model
This change adds a new ProjectBuildModel that can be used to represent a
whole project. A file cache is used within each ProjectBuildModel that
should reduce the amount of work needed to parse large projects. This
also has the advantage that all models created through the
ProjectBuildModel will have a consistent view of the world (Note: apart
from applied plugins, these are currently stored at the model level).

Bug: 74100881
Test: Unit tests added.
Change-Id: I6da0ed4a6cbc0c0cb499943ea956b504e1ec38bf