Initial error reporting for nav editor
Most of the changes are factoring the common bits of the existing error
panel into common. Notably, the types of issues are no longer hardcoded
in IssueModel, but instead are added as IssueProviders by the surface.

We don't have any nav-specific issues yet, but since the lint
IssueProvider is common it should be straightforward to enable common
lint checks or add new ones as needed.

Test: Existing tests updated.
Bug: 63356730
Change-Id: I5ee87ecdfd6b6e1228ff1692ad7dc0ff944c72df