Add new Java library module default name changed
When de-duplicating a directory name ("lib"), we get a warning on
Windows if the full path contains a space. We should not try generating
new names ("lib1", "lib2", ..., "lib100") as they will have the same
warning.
This CL updates the de-duplicating code to only check for ERRORS, and
ignore WARNINGS, INFO, etc.

Change-Id: Ie6e47f25087a211c416d6d7ef31fd9dafae001ec
Fixes: 74346753
Test: Added WizardUtilsTest