[AVD Mgr] Check the disk for AVDs before populating the UI list
In automated testing, an empty list of AVDs is sometimes
displayed, even though an AVD was created immediately prior.

In the DevicePicker constructor, the list of AVDs was populated
before the disk was checked. This resulted in an empty list. In
subsequent processing, the disk was checked and the list was
updated, so that initial empty list *should* never have been used.

Much of the list processing happens on separate threads--either
the UI thread (via invokeLater()) or the pooled thread (via
executeOnPooledThread()). Perhaps variations in this threading
sometimes made the initial list come last and be used in the UI
display.

This change removes the work from the DevicePicker constructor.
Now DeployTargetPickerDialog schedules the check on the pooled
thread and displays a spinner with "Looking for virtual devices"
until that check completes.

Fixes: 72748687
Bug:   74779369

Test: This addresses a failure in existing automated tests.
      DevicePickerListModelTest still passes.

Change-Id: I9f902a3c2ee89339b810f2d16621559ebb27e173