Introducing inspect trace mode to CpuProfilerStage
Add a boolean field, defaulted to false, to CpuProfilerStage and allow
setting it to true at construction time. This field makes the stage to
be initiated in a "Inspect Trace" mode, which will be used when
importing a CPU trace file. This mode can only be set if the Import
traces flag (introduced in this CL) is set.

Bug: 74341682
Test: CpuProfilerStageTest#inspectTraceModeOnlyEnabledWhenImportFlagIsSet
Change-Id: Ic96f3f4c2a18580d2579207626ea541981809daa