[Gradle Build Model] Add support for moving properties
This changes adds support for moving both basic statements and blocks.
We move statements by moving the whole PsiElement tree to its new location.

Note: Currently when calling applyChanges() on a model that has invoked
a call to GradleDslWriter#moveDslElement the model will be left in a invalid
state. Any other operations performed on the model are likely to fail. To
prevent this GradleBuildModel#reparse must be called. Currently there is
no way to invoke moveDslElement from the API package.

Test: Unit tests added.
Bug: N/A
Change-Id: Id54aca80408c327da6a05a29b6f6f1ee9cdfca60