Quick fix for NPE when a selected element is filtered.
A bug was filed where the tester selected a row that was then filtered.
The filter causes the tree to have fewer rows, however the mouseAdapter
maintains an index of the last hovered row. This row is then used to
trigger a repaint on that row. In the case of filtering the row no longer
exists.

Test: later
Bug: 74421577
Change-Id: Idaa8ddcf53d9588d7b13ef111371a43b35068832