[New PSD] One declared dependency object per scope.
In order to simplify handling of declared dependencies and simplify the UI
this CL changes instantiation of declared dependencies from one per
unique artifact id to one per statement in the configuration file.

The CL updates the corresponding tests and replaces some of the tests in
PsAndroidModuleTest with similar in DependencyManagementTest.

Bug: 74940013
Test: Tests updated
Change-Id: Icebfd1ef7db4c05db491df1746fbbdf1063c5991