Flatten includes hierarchy when there's one library
This was a regression in the enhanced C++ header files feature
that was caught by QA tests. When there's exaclty one C++
library in a module then the prior behavior was that we hide
that library's node in the Android project view. We the header
files feature turned on I was mistakenly not flattening.

Bug: None, I found it by looking at QA test run

Test: It is covered by QA test but (obviously) not by unittests.
The issue was in a pre-existing class NdkModuleNode that doesn't have
tests.

Change-Id: Ibad1047917e8d5d04d028c82eb71fd84d3f3ceda