Wait for project validation when importing Bazel project
We've been seeing flakes for BazelStartupSyncTest where the test
times out expecting the IDE frame to be visible when in fact the
import wizard is still validating the project view file.

This CL resolves the issue by adding an additional step to
BazelGuiTestProjectSystem#importProject, during which the test
explicitly waits for project validation to finish. This wait is
still checking for the availability of the IDE frame, but it
fails early with a more specific error message if there's an issue
during project validation.

Bug: 73037396
Test: BazelStartupSyncTest
      + AswbGotoDeclarationTest
      + Bazel tests in NlEditorTest
      + Forced a project view issue and verified that the
        wait terminates early with an appropriate error message.
Change-Id: Ieba985b2a69e616f830025b951543f16d447cebb