Fixing issue where state chart never cleared mouse position
The image diff comes from new we are initializing the default mouse position to -1
instead of 0. As such elements that push against the edge of the screen do not have
a mouse over. This should of been what was done initially but missed. This also
means that the mouse over StateChart code had a bug for some time.
Test: Coming
Change-Id: I209bbe7845792912d6d1b2c813267bd5a1957aed
Fixes: 76089898