[Gradle Build Model] Fix applied file merge order.
This CL shifts where we parse the applied file to when we
meet the apply element. This allow us to ensure that all
elements that are applied are inserted into the property
list in the same place there apply appears.

This also allows us to correctly resolve variables where
the same name is used in different files.

To do this we create an empty block element to merge elements into,
this allows us to keep all first level properties in the APPLIED state.

Bug: N/A
Test: Unit tests updated
Change-Id: I487686be7c5bb40a267f9ee94acac72c1eff8b26