Fix invalidation of PsiResourceItems
This is a modified version of commit
https://goto.google.com/android-sha1/65547bbe4b4d48e976da9a64841d2bf4f607cd0b
that is compatible with IntelliJ 181.2784.17. When a PsiResourceItem is
created from a PSI listener, creation of smart PSI pointers is done
asynchronously to avoid the "Smart pointers shouldn't be created during
PSI changes" error. Another difference from the above commit is that
a hard reference to the original PSI element is replaced by a soft
reference to prevent pinning PSI to memory.

When getting out of the dumb mode, PsiResourceItems inner XmlTag might
be invalidated. Since they might be cached, this will cause crashes
along the way.

This change fixes that invalidation by moving to use SmartPointers so
that the XmlTag can still be correctly retrieved after the file
invalidation since the file is still there, just invalid after the dumb
mode. A lot of the incrementality of the ResourceFolderRepository is
based on detecting changes based on the pointer to the XmlTag so added
a new method to check if the item was pointing at some point to that
tag.

Fixes: 73623886
Test: ResourceFolderRepositoryTest.testFileInvalidationAfterDumbMode
Change-Id: Ib44dd7f7e8e3bfbe60efff3b89f53101307db56c