[Gradle Build Model] Move GradleFileModelTestCase to use Parameterized test runner
This change converts GradleFileModelTestCase and all subclasses to
use JUnit4 to allow it to be run using JUnit4s Parameterized test runner.

This will allow us to easily run all the tests for Kotlin as well
in the future.

We also start to move test data from string literals to actual data files,
allowing us to swap them out for kotlin alternatives. AaptOptionsModelTest
is the only class currently moved.

Bug: N/A
Test: Unit tests updated, pass locally.
Change-Id: Ie8f65adab644b8d2fca04ab3c3dc680434dd43f0