Check server is running before trying to stop it
If the client fails to launch and the call to accept() times out, the
server tries to shut itself down. This can fail with an
UninitializedPropertyAccessException if a call to start() has not
succeeded previously (and would operate on old bogus values if it had).

Also decrease the connection timeout to something reasonable.

Bug: 77342247
Change-Id: Ieea555f8c46d2043f406f9b26e1426c892a4bfd1