Ensure KeepAliveThread#cancel is only called once
Previously, it was likely that cancel would be called twice, since
JUnitClientImpl#stop closes the connection and then calls cancel. The
KeepAliveThread calls cancel on itself if it detects the connection is
closed. This causes duplicate disposal errors when exiting the IDE.

This CL introduces synchronization in the cancel method so that it can
only be executed once.

Bug: 77483435
Test: RestartIdeTest
Change-Id: I336fbf5057eadeaadbae42af2ea8f9b5f2ba851a
Test-Strategy: ui.test.groups=TEST_FRAMEWORK