Updating the height of the threads view.
The spec outlines a 38 pixel high thread state with a 2
pixel blue border on selection, and a 1 pixel gap between states
This change adds a 2 pixel border on selection and a 3 pixel
gap. When selected 2 of the 3 pixels are used for the selection
border, while leaving a 1 pixel gap, this gives us enough space
to select each thread without the selection overlapping the row
above and below the currently selected item.

Test: N/A, Visually verified
Fixes: 75024019
Change-Id: I790699baa66e61ac6a7d52aed43c00ad82471a13