Always select the capture in import trace mode
If the user attempts to unselect the capture, we will
automatically select the entire range of the capture.
CpuProfilerStage is a good place to handle logic specific
to "import trace mode", which promotes this solution.

One alternative is to keep the current selection range in
such scenarios, not selecting the entire capture range.
However, that would require SelectionModel or CaptureModel
to be aware of "import trace mode", which isn't ideal.

This CL re-lands I92106d5f6308217453d0a7b750ca6fff37c28f89 because it
was accidentally reverted by I3f714fc323a02911be98a0aba57ae9bef3f98e06.

Change-Id: I5fcb319e2a29407437b2a17c18e200fe1874b123
Fixes: 77204101
Test: CpuProfilerStageTest#captureAlwaysSelectedInImportTraceMode