[Gradle Build Model] Rework GradleDslMethodCall internals
This change reworks how GradleDslMethodCall stores its arguments
moving it from a half-baked internal lists to using
a GradleDslExpressionList. This has the effect that method calls
can now store all types of GradleDslExpressions instead of just
GradleDslSimpleExpressions.

We also deal with the fallout, such as rewiring parsing and
writing to take into account this extra layer.

Bug: N/A
Test: Existing
Change-Id: I008acbef97495e120160b54e9e6641bef14db8e1